add permissions field, and validate delete permission

+++ app/controllers/Papers.scala
@@ -30,7 +30,7 @@ object Papers extends Controller {
         p match {
           case Some(p) =>
             tryOrError {
-              if (UsernameAuth.hasView(p.username, session)) {
+              if (UsernameAuth.canView(p.username, session)) {
                 val paperJson = Paper.model2json(p)
                 Ok(views.html.paper(JsonResult.jsonSuccess(paperJson)))
               } else {
@@ -81,7 +81,7 @@ object Papers extends Controller {
                 Future.successful(
                   JsonResult.error("Input is not a valid json"))
               case Some(json) =>
-                if (UsernameAuth.hasView(json getAsString Paper.username, session)) {
+                if (UsernameAuth.canView(json getAsString Paper.username, session)) {
                   val newPaper = Paper.json2model(json)
                   if (oldpaper._id != newPaper._id) {
 
@@ -149,7 +149,7 @@ object Papers extends Controller {
             } yield {
               val paperJson = Paper.model2json(
                   p.getOrElse(throw new Exception("Paper not found")))
-              if (UsernameAuth.hasView(p.get.username, username)) {
+              if (UsernameAuth.canView(p.get.username, username)) {
                 JsonResult.success(paperJson)
               } else {
                 JsonResult.noPermission
@@ -170,10 +170,18 @@ object Papers extends Controller {
         case Some(j) =>
           tryOrError {
             val paper = j \ "paper"
-            PaperDAO.save(Paper.json2model(paper), ow = true).map {
-              le =>
-                JsonResult.success("")
+            val p = Paper.json2model(paper)
+            if (UsernameAuth.canView(p.username, req.session)) {
+              PaperDAO.save(Paper.json2model(paper), ow = true).map {
+                le =>
+                  JsonResult.success("")
+              }
+            } else {
+              Future.successful(
+                JsonResult.noPermission
+              )
             }
+
           }
         case None =>
           Future.successful(JsonResult.error("Invalid json input"))
@@ -193,7 +201,7 @@ object Papers extends Controller {
                   case Some(paper) =>
                     val newId = Generator.oid()
                     val nowms = System.currentTimeMillis()
-                    if (UsernameAuth.hasView(paper.username, req.session.get("username"))) {
+                    if (UsernameAuth.canView(paper.username, req.session.get("username"))) {
                       val newPaper = paper.copy(
                         _id = newId,
                         created = nowms,
@@ -250,13 +258,19 @@ object Papers extends Controller {
               Json.obj(Paper._id -> paperId, Paper.username -> u))
               .getOrElse(Json.obj(Paper._id -> paperId))
             for {
-              a <- PaperDAO.findOne(q)
+              a <- PaperDAO.findOneModel(q)
               r <- {
                 a match {
-                  case Some(le) =>
-                    PaperDAO.remove(q).map {
-                      _ =>
-                        JsonResult.success("")
+                  case Some(p) =>
+                    if (UsernameAuth.canView(p.username, username)) {
+                      PaperDAO.remove(q).map {
+                        _ =>
+                          JsonResult.success("")
+                      }
+                    } else {
+                      Future.successful {
+                        JsonResult.noPermission
+                      }
                     }
                   case None =>
                     Future.successful(
